SLCORE-1220 SLCORE-1221 Refactor with event handling

+++ backend/telemetry/src/main/java/org/sonarsource/sonarlint/core/telemetry/measures/payload/TelemetryMeasuresBuilder.java
@@ -99,11 +99,11 @@ public class TelemetryMeasuresBuilder {
   private void addIssuesMeasures(List<TelemetryMeasuresValue> values) {
     var newIssuesFound = storage.getNewIssuesFoundCount();
     if (newIssuesFound > 0) {
-      values.add(new TelemetryMeasuresValue("ide_issues.found", Integer.toString(newIssuesFound), INTEGER, DAILY));
+      values.add(new TelemetryMeasuresValue("ide_issues.found", Long.toString(newIssuesFound), INTEGER, DAILY));
     }
     var issuesFixed = storage.getIssuesFixedCount();
     if (issuesFixed > 0) {
-      values.add(new TelemetryMeasuresValue("ide_issues.fixed", Integer.toString(issuesFixed), INTEGER, DAILY));
+      values.add(new TelemetryMeasuresValue("ide_issues.fixed", Long.toString(issuesFixed), INTEGER, DAILY));
     }
   }
 }
